moved dolt_docs tags to reserved tag space
