+++ go/libraries/doltcore/doltdb/root_val.go
@@ -40,10 +40,6 @@ const (
 	DocPkColumnName   = "doc_name"
 	DocTextColumnName = "doc_text"
 )
-const (
-	DocNameTag = iota
-	DocTextTag
-)
 
 // RootValue defines the structure used inside all Liquidata noms dbs
 type RootValue struct {
@@ -923,7 +919,7 @@ func addValuesToDocs(ctx context.Context, tbl *Table, sch *schema.Schema, docDet
 func AddValueToDocFromTbl(ctx context.Context, tbl *Table, sch *schema.Schema, docDetail DocDetails) (DocDetails, error) {
 	if tbl != nil && sch != nil {
 		pkTaggedVal := row.TaggedValues{
-			DocNameTag: types.String(docDetail.DocPk),
+			schema.DocNameTag: types.String(docDetail.DocPk),
 		}
 
 		docRow, ok, err := tbl.GetRowByPKVals(ctx, pkTaggedVal, *sch)
@@ -932,7 +928,7 @@ func AddValueToDocFromTbl(ctx context.Context, tbl *Table, sch *schema.Schema, d
 		}
 
 		if ok {
-			docValue, _ := docRow.GetColVal(DocTextTag)
+			docValue, _ := docRow.GetColVal(schema.DocTextTag)
 			docDetail.Value = docValue
 		} else {
 			docDetail.Value = nil
@@ -947,7 +943,7 @@ func AddValueToDocFromTbl(ctx context.Context, tbl *Table, sch *schema.Schema, d
 func AddNewerTextToDocFromTbl(ctx context.Context, tbl *Table, sch *schema.Schema, doc DocDetails) (DocDetails, error) {
 	if tbl != nil && sch != nil {
 		pkTaggedVal := row.TaggedValues{
-			DocNameTag: types.String(doc.DocPk),
+			schema.DocNameTag: types.String(doc.DocPk),
 		}
 
 		docRow, ok, err := tbl.GetRowByPKVals(ctx, pkTaggedVal, *sch)
@@ -955,7 +951,7 @@ func AddNewerTextToDocFromTbl(ctx context.Context, tbl *Table, sch *schema.Schem
 			return DocDetails{}, err
 		}
 		if ok {
-			docValue, _ := docRow.GetColVal(DocTextTag)
+			docValue, _ := docRow.GetColVal(schema.DocTextTag)
 			doc.NewerText = []byte(docValue.(types.String))
 		} else {
 			doc.NewerText = nil
@@ -967,7 +963,7 @@ func AddNewerTextToDocFromTbl(ctx context.Context, tbl *Table, sch *schema.Schem
 }
 
 func addNewerTextToDocFromRow(ctx context.Context, r row.Row, doc *DocDetails) (DocDetails, error) {
-	docValue, ok := r.GetColVal(DocTextTag)
+	docValue, ok := r.GetColVal(schema.DocTextTag)
 	if !ok {
 		doc.NewerText = nil
 	} else {
@@ -981,7 +977,7 @@ func addNewerTextToDocFromRow(ctx context.Context, r row.Row, doc *DocDetails) (
 }
 
 func addDocPKToDocFromRow(r row.Row, doc *DocDetails) (DocDetails, error) {
-	colVal, _ := r.GetColVal(DocNameTag)
+	colVal, _ := r.GetColVal(schema.DocNameTag)
 	if colVal == nil {
 		doc.DocPk = ""
 	} else {
