+++ go/libraries/doltcore/schema/column.go
@@ -16,15 +16,12 @@ package schema
 
 import (
 	"errors"
-	"math"
 	"strings"
 
 	"github.com/liquidata-inc/dolt/go/libraries/doltcore/schema/typeinfo"
 	"github.com/liquidata-inc/dolt/go/store/types"
 )
 
-// ReservedTagMin is the start of a range of tags which the user should not be able to use in their schemas.
-const ReservedTagMin uint64 = 1 << 63
 
 var (
 	// KindToLwrStr maps a noms kind to the kinds lowercased name
@@ -35,9 +32,6 @@ var (
 )
 
 var (
-	// InvalidTag is used as an invalid tag
-	InvalidTag uint64 = math.MaxUint64
-
 	// InvalidCol is a Column instance that is returned when there is nothing to return and can be tested against.
 	InvalidCol = Column{
 		"invalid",
